fix: add initial generation script
This creates a naive generator implementation that we can expand on, where each layer is defined by a file that consists of all available options. As we get more artwork we can revisit this structure to simplify.

This also adds a few npm packages to the genration folder, running "npm i" then "npx ts-node generate-svgs.ts" will run the script.

Everything is hardcoded for now but two samples were generated and included as an example.